TresPi/server/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1

numpy==1.26.4

faster-whisper
ctranslate2

coqui-tts

transformers==4.38.2

soundfile
librosa

groq
python-dotenv

websockets
fastapi
uvicorn[standard]

webrtcvad

num2words